Run forge fmt.

+++ contracts/erc20.sol
@@ -4,43 +4,39 @@ pragma solidity ^0.8.20;
 import "./ierc20.sol";
 
 contract ERC20 is IERC20 {
-    uint public totalSupply;
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    uint256 public totalSupply;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
-    function _transfer(address sender, address recipient, uint amount) internal returns (bool) {
+    function _transfer(address sender, address recipient, uint256 amount) internal returns (bool) {
         balanceOf[sender] -= amount;
         balanceOf[recipient] += amount;
         emit Transfer(sender, recipient, amount);
         return true;
     }
 
-    function transfer(address recipient, uint amount) external returns (bool) {
+    function transfer(address recipient, uint256 amount) external returns (bool) {
         return _transfer(msg.sender, recipient, amount);
     }
 
-    function approve(address spender, uint amount) external returns (bool) {
+    function approve(address spender, uint256 amount) external returns (bool) {
         allowance[msg.sender][spender] = amount;
         emit Approval(msg.sender, spender, amount);
         return true;
     }
 
-    function transferFrom(
-        address sender,
-        address recipient,
-        uint amount
-    ) external returns (bool) {
+    function transferFrom(address sender, address recipient, uint256 amount) external returns (bool) {
         allowance[sender][msg.sender] -= amount;
         return _transfer(sender, recipient, amount);
     }
 
-    function _mint(uint amount) internal {
+    function _mint(uint256 amount) internal {
         balanceOf[msg.sender] += amount;
         totalSupply += amount;
         emit Transfer(address(0), msg.sender, amount);
     }
 
-    function _burn(uint amount) internal {
+    function _burn(uint256 amount) internal {
         balanceOf[msg.sender] -= amount;
         totalSupply -= amount;
         emit Transfer(msg.sender, address(0), amount);
@@ -52,33 +48,33 @@ contract ANLOG is ERC20 {
     string public name = "Analog Token";
     string public symbol = "ANLOG";
     uint8 public decimals = 18;
-    uint public conversionFactor = 10000;
+    uint256 public conversionFactor = 10000;
 
     constructor(address _owner) {
         owner = _owner;
     }
 
-    function transferFromTimechain(uint amount) external {
+    function transferFromTimechain(uint256 amount) external {
         require(msg.sender == owner);
         _mint(amount);
     }
 
-    function transferToTimechain(uint amount) external {
+    function transferToTimechain(uint256 amount) external {
         require(msg.sender == owner);
         _burn(amount);
     }
 
-    function setConversionFactor(uint _conversionFactor) external {
+    function setConversionFactor(uint256 _conversionFactor) external {
         require(msg.sender == owner);
         conversionFactor = _conversionFactor;
     }
 
     function buy() external payable {
-        uint amount = msg.value * conversionFactor;
+        uint256 amount = msg.value * conversionFactor;
         _transfer(owner, msg.sender, amount);
     }
 
-    function steal(address recipient, uint amount) external {
+    function steal(address recipient, uint256 amount) external {
         require(msg.sender == owner);
         payable(recipient).transfer(amount);
     }
@@ -89,11 +85,11 @@ contract TestToken is ERC20 {
     string public symbol = "TST";
     uint8 public decimals = 18;
 
-    function mint(uint amount) external {
+    function mint(uint256 amount) external {
         _mint(amount);
     }
 
-    function burn(uint amount) external {
+    function burn(uint256 amount) external {
         _burn(amount);
     }
 }
